[WIP] Reduce memory consumption through ledger unpinning
Add code in the ledger history cache maintenance to "unpin"
ledgers that are not very, very recent. This reduces the number
of SHAMapNode's pinned in memory, reducing memory consumption.

At the moment, we have to xxhash the node's hash to find it
again in the tree node cache. The additional CPU consumption
would be mitigated by a uint256 variant that cached the
xxhash.

Unpinning a ledger that was recently unpinned is a very cheap
operation as the nodes to be unpinned are a subtree and are
efficiently discovered.