[EXPERIMENTAL] An even crazier idea
Inner nodes of unbacked SHAMap's still have strong pointers to their
children, but inner nodes of backed SHAMap's exclusively use weak
pointers to their children. When/if the nodes expire from cache, they
are removed from memory.

The thinking is that if we're using them, the cache will keep them in
memory. If we're not using them, we don't want them in memory. The use
of weak pointers reduces the cost of traversing the maps by eliminating
the need to find them in the cache.